requirements.txt
alabaster==0.7.12
alembic==1.0.2
anaconda-client==1.7.2
anaconda-navigator==1.9.2
anaconda-project==0.8.2
appdirs==1.4.3
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.24.0
aspy.yaml==1.1.1
astroid==2.0.4
astropy==3.0.5
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.4
beautifulsoup4==4.6.3
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==3.0.2
bokeh==1.0.0
boto==2.49.0
Bottleneck==1.2.1
cached-property==1.5.1
certifi==2018.10.15
cffi==1.11.5
cfgv==1.1.0
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
clyent==1.2.2
colorama==0.4.0
conda==4.5.11
conda-build==3.16.1
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.3.1
cx-Oracle==7.0.0
cycler==0.10.0
cymem==2.0.2
Cython==0.29
cytoolz==0.9.0.1
dask==0.19.4
datashape==0.5.4
decorator==4.3.0
defusedxml==0.5.0
diff-match-patch==20121119
dill==0.2.8.2
distributed==1.23.3
dj-database-url==0.5.0
Django==2.1.3
django-excel==0.0.10
django-heroku==0.3.1
django-import-export==1.1.0
djangorestframework==3.9.0
docutils==0.14
en-core-web-md==2.0.0
en-core-web-sm==2.0.0
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.9
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Login==0.4.1
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
futures==3.1.1
fuzzywuzzy==0.17.0
gevent==1.3.7
glob2==0.6
gmpy2==2.0.8
googletrans==2.3.0
goslate==1.5.1
greenlet==0.4.15
gunicorn==19.9.0
h5py==2.8.0
heapdict==1.0.0
html5lib==1.0.1
hyperlink==18.0.0
identify==1.1.7
idna==2.7
imageio==2.4.1
imagesize==1.1.0
imbalanced-learn==0.4.3
imblearn==0.0
importlib-metadata==0.6
importlib-resources==1.0.1
incremental==17.5.0
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
ItsDangerous==1.0.0
jdcal==1.4
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.2
jupyterlab-launcher==0.13.1
jupyterlab-server==0.2.0
keyring==15.1.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
libarchive-c==2.8
llvmlite==0.25.0
lml==0.0.4
locket==0.2.0
lxml==4.2.5
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.0.0
mccabe==0.6.1
missingno==0.4.1
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
more-itertools==4.3.0
mpmath==1.0.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
multipledispatch==0.6.0
murmurhash==1.0.1
navigator-updater==0.2.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
nltk==3.3
nodeenv==1.3.2
nose==1.3.7
notebook==5.7.0
numba==0.40.0
numexpr==2.6.8
numpy==1.15.3
numpydoc==0.8.0
odfpy==1.3.5
odo==0.5.1
olefile==0.46
openpyxl==2.5.9
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.9
path.py==11.5.0
pathlib2==2.3.2
patsy==0.5.0
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pkginfo==1.4.2
plac==0.9.6
plotly==3.4.0
pluggy==0.8.0
ply==3.11
pre-commit==1.12.0
preshed==2.0.1
prometheus-client==0.4.2
prompt-toolkit==2.0.6
psutil==5.4.7
psycopg2==2.7.5
ptyprocess==0.6.0
py==1.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pyexcel==0.5.9.1
pyexcel-io==0.5.9.1
pyexcel-ods==0.5.3
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.6
pyflakes==2.0.0
Pygments==2.2.0
pylint==2.1.1
pyodbc==4.0.24
pyOpenSSL==18.0.0
pyparsing==2.2.2
PySocks==1.6.8
pytest==3.9.1
pytest-arraydiff==0.2
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.3.0
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.5.1
qtconsole==4.4.2
QtPy==1.5.2
regex==2018.1.10
requests==2.19.1
retrying==1.3.3
rope==0.11.0
ruamel-yaml==0.15.46
scikit-image==0.14.0
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.0.5
spacy==2.0.16
Sphinx==1.8.1
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
SQLAlchemy==1.2.12
statsmodels==0.9.0
stop-words==2018.7.23
sympy==1.3
tables==3.4.4
tablib==0.12.1
tblib==1.3.2
terminado==0.8.1
testpath==0.4.2
texttable==1.4.0
thinc==6.12.0
toml==0.10.0
toolz==0.9.0
tornado==5.1.1
tox==3.5.2
tqdm==4.26.0
traitlets==4.3.2
translate==3.5.0
Twisted==17.5.0
typed-ast==1.1.0
ujson==1.35
unicodecsv==0.14.1
urllib3==1.23
virtualenv==16.0.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==4.1
widgetsnbextension==3.4.2
wordninja==0.1.5
wrapt==1.10.11
WTForms==2.2.1
xlrd==1.1.0
XlsxWriter==1.1.2
xlwings==0.12.1
xlwt==1.2.0
zict==0.1.3
zope.interface==4.5.0
